refactor: propose multisig transactions
