+++ apps/extension/src/app/features/rpc-stacks-transaction-request/stacks/use-propose-stacks-transaction.ts
@@ -2,7 +2,9 @@ import { useCallback } from 'react';
 import { useLocation, useNavigate } from 'react-router';
 
 import { type StacksTransactionWire } from '@stacks/transactions';
+import { useMutation } from '@tanstack/react-query';
 
+import { createProposeMultisigTransactionMutationConfig } from '@leather.io/queries';
 import { type RpcMethodNames, createRpcSuccessResponse } from '@leather.io/rpc';
 import { deriveStxMultisigAddress } from '@leather.io/stacks';
 import { delay, isString } from '@leather.io/utils';
@@ -13,7 +15,7 @@ import { analytics } from '@shared/utils/analytics';
 
 import { useRpcRequestParams } from '@app/common/hooks/use-rpc-request-params';
 import { getPolicyAuthNetworkId } from '@app/features/multisig/multisig-network';
-import { useProposeMultisigTransaction } from '@app/features/multisig/use-propose-multisig-transaction';
+import { useSignProposalCommitment } from '@app/features/multisig/use-sign-proposal-commitment';
 import { useCurrentStacksAccount } from '@app/store/accounts/blockchain/stacks/stacks-account.hooks';
 import { useCurrentNetwork } from '@app/store/networks/networks.selectors';
 import { useCurrentPolicy } from '@app/store/policy/policy.selectors';
@@ -40,7 +42,10 @@ export function useProposeStacksTransaction(method: RpcMethodNames) {
   const network = useCurrentNetwork();
   const navigate = useNavigate();
   const { pathname } = useLocation();
-  const { mutateAsync: proposeMultisigTransaction } = useProposeMultisigTransaction();
+  const signProposalCommitment = useSignProposalCommitment();
+  const { mutateAsync: proposeMultisigTransaction } = useMutation(
+    createProposeMultisigTransactionMutationConfig({ signProposalCommitment })
+  );
 
   return useCallback(
     async (unsignedTx: StacksTransactionWire) => {
